Update chat_bot.py- comment fix and adding to code
Here I added a statement to print in a piece of code. In this section, when patient is prompted to put the days they have been experiencing symptoms, if it is more than 13 days they are advised to seek consultation from a doctor. I added in a statement about our main three for this project. Those main three being: Primary Care, Urgent Care, or ER. I also made a minor fix in the code.